anim manager migration modifications (tests not passing yet)
git-svn-id: file:///home/hoijui/src/IsaViz/tmp2/zvtm.svn/code@4646 74e165f3-c697-40a6-91af-552c8936d7fe

+++ zvtm-core/branches/TRY-RP-MigrateAnimFramework/src/main/java/fr/inria/zvtm/animation/DefaultTimingHandler.java
@@ -25,6 +25,6 @@ public abstract class DefaultTimingHandler implements TimingHandler{
     public void begin(Object subject, Animation.Dimension dim){}
     public void end(Object subject, Animation.Dimension dim){}
     public void repeat(Object subject, Animation.Dimension dim){}
-    public abstract void timingEvent(float fraction, 
+    public abstract void timingEvent(double fraction, 
 				     Object subject, Animation.Dimension dim);
 }
